создали подключаемый шаблон form.html через include везде где формы вставили реализовали удаление подправили редактирование

+++ source/To_do_list/views.py
@@ -1,5 +1,9 @@
 from django.shortcuts import render, redirect, get_object_or_404
+
+from To_do_list.forms import TaskForm
 from To_do_list.models import Task, status_choices
+
+
 # Create your views here.
 
 
@@ -18,34 +22,51 @@ def task_view(request, pk):
 
 def task_create_view(request):
     if request.method == 'GET':
-        return render(request, 'task_create.html', context={'status': status_choices})
+        form = TaskForm()
+        return render(request, 'task_create.html', context={'form': form})
     elif request.method == 'POST':
-        title = request.POST.get('title')
-        status = request.POST.get('status')
-        date = request.POST.get('date')
-        if not date:
-            date = None
-        task = Task.objects.create(title=title, status=status, date=date)
-
+        form = TaskForm(data=request.POST)
+        if form.is_valid():
+            task = Task(
+                title=form.cleaned_data.get('title'),
+                status=form.cleaned_data.get('status'),
+                date=form.cleaned_data.get('date'),
+                add_info=form.cleaned_data.get('add_info')
+            )
+            task.save()
+        else:
+            return render(request, 'task_create.html', context={'form': form})
         return redirect('task_view', pk=task.id)
 
 
 def task_update_view(request, pk):
-    task = get_object_or_404(Task, id=pk)
+    task = get_object_or_404(Task, pk=pk)
     if request.method == 'GET':
-        return render(request, 'task_update.html', context={'task': task})
+        form = TaskForm(initial={
+            "title": task.title,
+            "status": task.status,
+            "date": task.date,
+            "add_info": task.add_info,
+        })
+        return render(request, 'task_update.html', context={'form': form, "id": task.id})
     elif request.method == 'POST':
-        task.title = request.POST.get('title')
-        task.status = request.POST.get('status')
-        task.date = request.POST.get('date')
-        task.save()
+        form = TaskForm(data=request.POST)
+        if form.is_valid():
+            task.title = form.cleaned_data.get('title')
+            task.status = form.cleaned_data.get('status')
+            task.date = form.cleaned_data.get('date')
+            task.add_info = form.cleaned_data.get('add_info')
+
+            task.save()
+        else:
+            return render(request, 'task_update.html', context={'form': form, "id": task.id})
         return redirect('task_view', pk=task.id)
 
 
-
 def task_delete_view(request, pk):
-    task = get_object_or_404(Task, id=pk)
-    task.delete()
-    return redirect('index')
-
-
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'GET':
+        return render(request, 'task_delete.html', context={'task': task})
+    elif request.method == 'POST':
+        task.delete()
+        return redirect('index')
